% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{popsimB}
\alias{popsimB}
\title{Simulation time series data for population B}
\format{
An object of class \code{array} of dimension 50 x 10 x 15.
}
\usage{
popsimB
}
\description{
A dataset containing values of 10 interested
variables of 15 subjects over 50 periods.
}
\seealso{
\code{\link{popsimA}}.
}
\keyword{datasets}
